Move cleanupTempVolumes and cleanupRunMounts to run_common.go
